Fix for wrong chdir by adding needed path separator
Because source_dir doesn't end in / when importing a file from D:\ (on Windows), later in assimp.rb 'cd /d "D:"' is being done (should be 'cd /d "D:/"'), which results in just keeping the current directory on D: because it has no path, just a drive letter, which results in the current directory of the drive not being changed (stays something like /Users/username/Documents for instance), and not the root / as expected.
The File.join statements are smart enough to properly join their first arguments ending on File::SEPARATOR with their second arguments.